requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
pytesseract
motor
pydantic
